Adversary made into Imposter
Added adversary to recipients for users and randomized the names along with giving each recipient a specific color.

moved internal css code in message html to main.css and gave it class "gray-blue"

added db field for am/pm so users can specify  when they want to meet.

added location and time to adversary page and changed a few colors on the page so it's more inline with the website.